+++ src/components/Plots/PlotTypes/Histogram.tsx
@@ -6,6 +6,8 @@ import PlotTypeProps from '../../../types/plottypeprops.interface';
 import VegaDataPlot from '../VegaDataPlot';
 import { useAppSelector } from '../../../app/store';
 import { selectProjectMetadataFields } from '../../../app/projectMetadataSlice';
+import { ColorSchemeSelectorPlotStyle } from '../../Trees/TreeControls/SchemeSelector';
+import { useStateFromSearchParamsForPrimitive } from '../../../utilities/helperUtils';
 
 // We will check for these in order in the given dataset, and use the first found as default
 // Possible enhancement: allow preferred field to be specified in the database, overriding these
@@ -34,12 +36,34 @@ function Histogram(props: PlotTypeProps) {
   const { fields, fieldUniqueValues } = useAppSelector(
     state => selectProjectMetadataFields(state, plot?.projectAbbreviation),
   );
+  const searchParams = new URLSearchParams(window.location.search);
   const [categoricalFields, setCategoricalFields] = useState<string[]>([]);
   const [numericFields, setNumericFields] = useState<string[]>([]);
-  const [xAxisField, setXAxisField] = useState<string>('');
-  const [colourField, setColourField] = useState<string>('none');
-  const [binMode, setBinMode] = useState<string>('auto');
-  const [stepSize, setStepSize] = useState<number>(1);
+  const [xAxisField, setXAxisField] = useStateFromSearchParamsForPrimitive<string>(
+    'xAxisField',
+    '',
+    searchParams,
+  );
+  const [colourField, setColourField] = useStateFromSearchParamsForPrimitive<string>(
+    'colourField',
+    'none',
+    searchParams,
+  );
+  const [colourScheme, setColourScheme] = useStateFromSearchParamsForPrimitive<string>(
+    'colourScheme',
+    'spectral',
+    searchParams,
+  );
+  const [binMode, setBinMode] = useStateFromSearchParamsForPrimitive<string>(
+    'binMode',
+    'auto',
+    searchParams,
+  );
+  const [stepSize, setStepSize] = useStateFromSearchParamsForPrimitive<number>(
+    'stepSize',
+    1,
+    searchParams,
+  );
 
   // Set spec on load
   useEffect(() => {
@@ -68,8 +92,13 @@ function Histogram(props: PlotTypeProps) {
       if (localNumericFields.length === 0) {
         setPlotErrorMsg('No numeric fields found in project, cannot render plot');
       }
-      setXAxisField(getStartingField(preferredXAxisFields, localNumericFields));
+      if (xAxisField === '') {
+        setXAxisField(getStartingField(preferredXAxisFields, localNumericFields));
+      } else if (!localNumericFields.includes(xAxisField)) {
+        setPlotErrorMsg('Invalid field in URL, cannot render plot');
+      }
     }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [fields, setPlotErrorMsg]);
 
   useEffect(() => {
@@ -83,12 +112,17 @@ function Histogram(props: PlotTypeProps) {
 
   useEffect(() => {
     const setColorInSpec = (oldSpec: TopLevelSpec | null): TopLevelSpec | null =>
-      setColorInSpecToValue(oldSpec, colourField, fieldUniqueValues![colourField] ?? []);
+      setColorInSpecToValue(
+        oldSpec,
+        colourField,
+        fieldUniqueValues![colourField] ?? [],
+        colourScheme,
+      );
 
     if (fieldUniqueValues) {
       setSpec(setColorInSpec);
     }
-  }, [colourField, fieldUniqueValues]);
+  }, [colourField, colourScheme, fieldUniqueValues]);
 
   useEffect(() => {
     const addBinningToSpec = (oldSpec: TopLevelSpec | null): TopLevelSpec | null => {
@@ -120,7 +154,7 @@ function Histogram(props: PlotTypeProps) {
           onChange={(e) => setXAxisField(e.target.value)}
         >
           {
-            numericFields.map(field => <MenuItem value={field}>{field}</MenuItem>)
+            numericFields.map(field => <MenuItem key={field} value={field}>{field}</MenuItem>)
           }
         </Select>
       </FormControl>
@@ -135,10 +169,16 @@ function Histogram(props: PlotTypeProps) {
         >
           <MenuItem value="none">None</MenuItem>
           {
-            categoricalFields.map(field => <MenuItem value={field}>{field}</MenuItem>)
+            categoricalFields.map(field => <MenuItem key={field} value={field}>{field}</MenuItem>)
           }
         </Select>
       </FormControl>
+      {colourField !== 'none' && (
+        <ColorSchemeSelectorPlotStyle
+          selectedScheme={colourScheme}
+          onColourChange={(newColor) => setColourScheme(newColor)}
+        />
+      )}
       <FormControl size="small" sx={{ marginX: 1, marginTop: 1 }}>
         <InputLabel id="binning-auto-select-label">Binning</InputLabel>
         <Select